// Frontend/src/app/core/services/aminety-service/models/CreateAmenityRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateAmenityRequestDto = {
    name?: string | null;
    slotTypeId?: string;
    statusId?: string;
    location?: string | null;
    rules?: string | null;
    imageUrl?: string | null;
};